possible fix for ERP-306

+++ grails-app/controllers/groups/GroupClientProfileController.groovy
@@ -186,7 +186,7 @@ class GroupClientProfileController {
       profile {
         if (params.city)
           ilike('currentCity', "%" + params.city + "%")
-        if (params.int('gender') > 0)
+        if (params.byte('gender') > 0)
           eq('gender', params.byte('gender'))
         if (params.job && params.job.toInteger() > 0) {
           if (params.job.toInteger() == 1)
